+++ Expanded Cheat Menu/Code/Menus/HelpFunc.lua
@@ -125,7 +125,8 @@ do -- ModUpload
 
 			-- add some text to ECM description to hopefully reduce people reporting the mod.
 			if mod.id == ChoGGi.id then
-				mod.description = Strings[302535920000990--[["You need to have a mouse to use this mod."]]] .. Strings[302535920000887--[["If you have any issues with this mod, please send me a bug report instead of reporting the mod.
+				mod.description = Strings[302535920000990--[["You need to have a mouse to use this mod."]]] .. "\n"
+					.. Strings[302535920000887--[["If you have any issues with this mod, please send me a bug report instead of reporting the mod.
 You can contact me through:
 Github: https://github.com/ChoGGi/SurvivingMars_CheatMods
 Discord: ChoGGi#9210
@@ -265,8 +266,7 @@ email: ECM@choggi.org"]]] .. "\n\n\n" .. mod.description
 			if not mod.last_changes or mod.last_changes == "" then
 				local version = mod.version_major .. "." .. mod.version_minor
 				if testing then
-					mod.last_changes = Strings[302535920001428--[[v]]] .. version
-						.. "\nhttps://github.com/ChoGGi/SurvivingMars_CheatMods/tree/master/Mods%20ChoGGi/"
+					mod.last_changes = "https://github.com/ChoGGi/SurvivingMars_CheatMods/tree/master/Mods%20ChoGGi/"
 						.. (orig_title or mod.title):gsub(" ","%%20") .. "/changes.txt"
 				else
 					mod.last_changes = version