+++ Expanded Cheat Menu/Code/Testing.lua
@@ -11,25 +11,28 @@ end
 
 
 ----------------------
-print("remove next update after tito")
-local Colonist = Colonist
-local funcs = {
-	"LogStatClear",
-	"AddToLog",
-}
-for i = 1, #funcs do
-	local func = Colonist[funcs[i]]
-	Colonist[funcs[i]] = function(self, log, ...)
-		if log then
-			return func(self, log, ...)
-		end
-	end
-end
+
+--~ if LuaRevision <= 1001514 then
+--~ local Colonist = Colonist
+--~ local funcs = {
+--~ 	"LogStatClear",
+--~ 	"AddToLog",
+--~ }
+--~ for i = 1, #funcs do
+--~ 	local func = Colonist[funcs[i]]
+--~ 	Colonist[funcs[i]] = function(self, log, ...)
+--~ 		if log then
+--~ 			return func(self, log, ...)
+--~ 		end
+--~ 	end
+--~ end
+
 ----------------------
 --~ function IsDevelopmentSandbox()
 --~ 	return true
 --~ end
 
+--~ end
 
 
 -- load up onmsg prints