+++ Mods ChoGGi/Applicants Age/Code/Script.lua
@@ -22,10 +22,18 @@ function OnMsg.NewDay() -- NewSol...
 		return
 	end
 
+	local ColonistAgeGroups = const.ColonistAgeGroups
 	local g_ApplicantPool = g_ApplicantPool
 	for i = 1, #g_ApplicantPool do
 		local applicant = g_ApplicantPool[i][1]
 		applicant.age = applicant.age + 1
+		-- bump age_trait
+		local agegroup = ColonistAgeGroups[applicant.age_trait]
+		if applicant.age > agegroup.min then
+			if agegroup.next_agegroup then
+				applicant.age_trait = agegroup.next_agegroup
+			end
+		end
 	end
 
 end