+++ Mods ChoGGi/Construction Show Max Radius Range/Code/Script.lua
@@ -32,7 +32,7 @@ local function AddRadius(self, radius)
 	self:Attach(circle)
 end
 
-local cls_saved_settings = {"TriboelectricScrubber", "SubsurfaceHeater", "CoreHeatConvector", "ForestationPlant"}
+local cls_saved_settings = {"ChoGGi_TriboelectricSensorTower", "TriboelectricScrubber", "SubsurfaceHeater", "CoreHeatConvector", "ForestationPlant"}
 local cls_heaters = {"SubsurfaceHeater", "CoreHeatConvector"}
 
 local orig_CursorBuilding_GameInit = CursorBuilding.GameInit